tools/requirements.txt
pillow
PyMuPDF
opencv-python
numpy
modelscope
huggingface_hub